server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Gabriel <gabrielfgsantos95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
bcrypt = "^4.1.3"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev src/main.py'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'

[tool.pytest.ini_options]
pythonpath= "."
addopts = "-p no:warnings"

[tool.coverage.run]
omit = ["__init__.py"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
